final update

+++ src/labs_examples/datatypes_operators/labs/Exercise_07.java
@@ -21,7 +21,7 @@ public class Exercise_07 {
         // prompt user
         System.out.print("Enter a number in days between 1 and 1,000,000: ");
         // assign input to variable as int
-        int days = scanner.nextInt();
+        long days = scanner.nextInt();
 
         // write completed code here
         long seconds = days * 24 * 60 * 60;
